requirements.txt
streamlit
pandas
python-dateutil
faiss-cpu
numpy
openpyxl
ollama
requests
tqdm